requirements.txt
Django>=3.2
psycopg2-binary>=2.8.0
PyYAML>=5.4.0

# Test dependencies
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
model-bakery>=1.20.0
